refactor(API): Remove API calls from ThreadController
